requirements.txt
alembic==1.2.1
aniso8601==8.0.0
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coverage==4.5.4
coveralls==1.8.2
docopt==0.6.2
factory-boy==2.12.0
Faker==2.0.3
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
psycopg2-binary==2.7.7
py==1.8.0
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.2.1
pytest-cov==2.8.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.10
text-unidecode==1.3
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
